Added WP color scheme to M3Theme
