+++ WordPress/src/main/java/org/wordpress/android/ui/main/feedbackform/FeedbackFormScreen.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.wordpress.android.R
+import org.wordpress.android.ui.compose.theme.M3Theme
 
 @Composable
 fun FeedbackFormScreen(
@@ -129,7 +130,7 @@ private fun Screen(
     content: @Composable () -> Unit,
     onCloseClick: () -> Unit
 ) {
-    AppThemeMaterial3 {
+    M3Theme {
         Scaffold(
             topBar = {
                 TopAppBar(
